avancement nextCompetitor + arbitrage

+++ app/src/main/java/com/example/androidprojetparkour/api/models/performancesObstacles/PerformanceObstaclesItem.kt
@@ -6,7 +6,7 @@ data class PerformanceObstaclesItem(
     val id: Int,
     val obstacle_id: Int,
     val performance_id: Int,
-    val time: Long,
+    val time: Int,
     val to_verify: Int,
     val updated_at: String
 )
